Improve textual display of keys of type CompositeData.  Perviously fixed for CJSON, now for text.
