+++ src/main/scala/cjmx/cli/MessageFormatter.scala
@@ -2,7 +2,6 @@ package cjmx.cli
 
 import javax.management.{ObjectName, Attribute, MBeanInfo}
 
-
 trait MessageFormatter {
   def formatNames(names: Seq[ObjectName]): Seq[String]
   def formatAttributes(attrs: Seq[(ObjectName, Seq[Attribute])]): Seq[String]
